requirements.txt
flask
flask-cors
azure-ai-textanalytics